Legacy Widget: Move useForm() up a level
